Puts UITextField delegate methods of LoginViewController into extension, Sets expected byte sizes of returned JSON objects into constants
